Get error about DynamicSupervisor was given as a child to a supervisor
** (Mix) Could not start application kafka_test: KafkaTest.Application.start(:normal, []) returned an error: shutdown: failed to start child: SigstrKafkaMonitor
    ** (EXIT) an exception was raised:
        ** (ArgumentError) The module DynamicSupervisor was given as a child to a supervisor
but it does not implement child_spec/1.

If you own the given module, please define a child_spec/1 function
that receives an argument and returns a child specification as a map.
For example:

    def child_spec(opts) do
      %{
        id: __MODULE__,
        start: {__MODULE__, :start_link, [opts]},
        type: :worker,
        restart: :permanent,
        shutdown: 500
      }
    end

Note that "use Agent", "use GenServer" and so on automatically define
this function for you.

However, if you don't own the given module and it doesn't implement
child_spec/1, instead of passing the module name directly as a supervisor
child, you will have to pass a child specification as a map:

    %{
      id: DynamicSupervisor,
      start: {DynamicSupervisor, :start_link, [arg1, arg2]}
    }